refactor(operand): remove unused code
