+++ packages/app-builder/src/components/Scenario/AstBuilder/AstBuilderNode/Operand/OperandEditor/OperandEditor.tsx
@@ -1,4 +1,5 @@
 import {
+  allAggregators,
   type AstNode,
   isDatabaseAccess,
   isPayload,
@@ -14,7 +15,6 @@ import {
 } from '@app-builder/models';
 import { newTimeAddLabelledAst } from '@app-builder/models/LabelledAst/TimeAdd';
 import { newTimeNowLabelledAst } from '@app-builder/models/LabelledAst/TimeNow';
-import { allAggregators } from '@app-builder/services/editor';
 import {
   adaptAstNodeFromEditorViewModel,
   adaptEditorNodeViewModel,
@@ -147,19 +147,17 @@ function OperandEditorContent({
 }) {
   const { t } = useTranslation('scenarios');
   const options = useMemo(() => {
-    const databaseAccessors = builder.input.identifiers.databaseAccessors.map(
-      (node) =>
-        newDatabaseAccessorsLabelledAst({
-          dataModel: builder.input.dataModel,
-          node,
-        }),
+    const databaseAccessors = builder.input.databaseAccessors.map((node) =>
+      newDatabaseAccessorsLabelledAst({
+        dataModel: builder.input.dataModel,
+        node,
+      }),
     );
-    const payloadAccessors = builder.input.identifiers.payloadAccessors.map(
-      (node) =>
-        newPayloadAccessorsLabelledAst({
-          triggerObjectTable: builder.input.triggerObjectTable,
-          node,
-        }),
+    const payloadAccessors = builder.input.payloadAccessors.map((node) =>
+      newPayloadAccessorsLabelledAst({
+        triggerObjectTable: builder.input.triggerObjectTable,
+        node,
+      }),
     );
     const customLists = builder.input.customLists.map(newCustomListLabelledAst);
     const functions = [
