+++ packages/app-builder/src/models/LabelledAst/LabelledAst.ts
@@ -12,15 +12,10 @@ import {
   isUndefinedAstNode,
 } from '../ast-node';
 import { type DataType, type EnumValue, type TableModel } from '../data-model';
-import {
-  type EditorIdentifiersByType,
-  getIdentifiersFromAstNode,
-} from '../identifier';
 import { newAggregatorLabelledAst } from './Aggregator';
 import { newConstantLabelledAst, newEnumConstantLabelledAst } from './Constant';
 import { newCustomListLabelledAst } from './CustomList';
 import { newDatabaseAccessorsLabelledAst } from './DatabaseAccessors';
-import { getAstNodeDisplayName } from './getAstNodeDisplayName';
 import { newPayloadAccessorsLabelledAst } from './PayloadAccessor';
 import { newTimeAddLabelledAst } from './TimeAdd';
 import { newTimeNowLabelledAst } from './TimeNow';
@@ -29,7 +24,6 @@ import { newUndefinedLabelledAst } from './Undefined';
 //TODO(combobox): find a better naming
 export interface LabelledAst {
   name: string;
-  description?: string;
   operandType:
     | 'Constant'
     | 'CustomList'
@@ -39,6 +33,9 @@ export interface LabelledAst {
     | 'Undefined'
     | 'unknown';
   dataType: DataType;
+
+  // Specific to the OperandDescription, move it as conditional type
+  description?: string;
   astNode: AstNode;
   values?: EnumValue[];
 }
@@ -106,35 +103,4 @@ export function adaptLabelledAst(
   }
 
   return null;
-}
-
-/**
- * @deprecated Only used in Scenario/Formula/*
- */
-export function adaptLabelledAstFromAllIdentifiers(
-  astNode: AstNode,
-  identifiers: EditorIdentifiersByType,
-): LabelledAst {
-  const identifier = getIdentifiersFromAstNode(astNode, identifiers);
-  if (identifier) {
-    return adaptLabelledAstFromIdentifier(identifier);
-  }
-  return {
-    name: getAstNodeDisplayName(astNode),
-    dataType: 'unknown',
-    operandType: 'unknown',
-    astNode,
-  };
-}
-
-/**
- * @deprecated Only used adaptLabelledAstFromAllIdentifiers
- */
-function adaptLabelledAstFromIdentifier(identifier: AstNode): LabelledAst {
-  return {
-    name: getAstNodeDisplayName(identifier),
-    dataType: 'unknown',
-    operandType: 'unknown',
-    astNode: identifier,
-  };
 }
