+++ packages/app-builder/src/models/index.ts
@@ -3,10 +3,8 @@ export * from './ast-operators';
 export * from './auth-errors';
 export * from './data-model';
 export * from './http-errors';
-export * from './identifier';
 export * from './LabelledAst';
 export * from './marble-session';
-export * from './operators';
 export * from './scenario';
 export * from './scenario-validation';
 export * from './user';
