+++ packages/app-builder/src/services/editor/index.ts
@@ -2,5 +2,4 @@ export * from './ast-expression';
 export * from './coerceToConstantsLabelledAst';
 export * from './current-scenario-iteration';
 export * from './editor-mode';
-export * from './identifiers';
 export * from './operators';
